+++ src/app/resources/floor/controller.js
@@ -1,5 +1,5 @@
 import FloorService from "./service";
-
+import Sector from "../../models/Sector";
 class FloorController {
   async list(req, res) {
     const { shelfId } = req.query;
@@ -28,7 +28,19 @@ class FloorController {
 
     const floorService = new FloorService();
 
+    const sectors = await Sector.findAll({
+      where: {
+        floorId: id,
+      },
+    });
+
+    if (sectors && sectors.length > 0)
+      return res.status(400).json({
+        message: "Não é possível excluir andar com setor incluído",
+      });
+
     const floor = await floorService.delete(id);
+
     if (!floor) return res.status(400).json({ message: "Andar não existe" });
 
     return res.sendStatus(204);
